+++ packages/browser/types/index.d.ts
@@ -1,4 +1,4 @@
 import { ResponseHandler } from "@hickory/root";
 import { BrowserHistoryOptions, BrowserHistory } from "./types";
 export * from "./types";
-export declare function Browser(fn: ResponseHandler, options?: BrowserHistoryOptions): BrowserHistory;
+export declare function browser(fn: ResponseHandler, options?: BrowserHistoryOptions): BrowserHistory;