+++ packages/in-memory/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./types";
 
-import { InMemory } from "./InMemory";
+import { in_memory } from "./in_memory";
 import { create_server_history } from "./create_server_history";
 
-export { InMemory, create_server_history };
+export { in_memory, create_server_history };